refactor(protocol): split receipt.ts to keep under 1500-LOC file-size cap
Validator hardening + cross-field invariants pushed receipt.ts to 1555 LOC
(1500 hard cap per scripts/check-file-size.sh). Splits cleanly along
existing seams:

- `receipt-types.ts` (214 LOC) — interface types + brand constructors
- `receipt-validator.ts` (~700 LOC) — validateReceipt + per-type key sets
- `receipt-utils.ts` (66 LOC) — pointer/isRecord/hasOwn/addError/etc.
- `receipt.ts` (~700 LOC) — codec only (receiptToJson + receiptFromJson)

receipt.ts re-exports the validator + types so consumer imports stay stable.
74/74 tests pass; tsc + biome clean.
